Moved engines comparison to benchmarking module; Extracted common logic in Measuring trait
